+++ internal/app/product/repo/read_model.go
@@ -214,12 +214,12 @@ func (rm *ReadModelImpl) dataToDTO(data *m_product.Data, now time.Time) (*contra
 	// Handle discount
 	if data.DiscountPercent.Valid {
 		discount, err := domain.NewDiscount(
-			data.DiscountPercent.Int64,
+			data.DiscountPercent.Float64, // Changed from Int64 to Float64
 			data.DiscountStartDate.Time,
 			data.DiscountEndDate.Time,
 		)
 		if err == nil && discount.IsValidAt(now) {
-			dto.DiscountPercent = &data.DiscountPercent.Int64
+			dto.DiscountPercent = &data.DiscountPercent.Float64 // Changed from Int64 to Float64
 			dto.DiscountActive = true
 			effectivePrice := discount.Apply(basePrice)
 			effectivePriceFloat, _ := effectivePrice.Float64()
